middleware-docker/requirement2.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==4.2.16
django-bootstrap-v5==1.0.11
django-bootstrap5==24.2
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
